Move `ask_for_local_branch` to `base_commands.py`
